UPD product forms price is now a float

+++ app/components/product card/admin product card/AdminProductCardEdit.tsx
@@ -131,7 +131,7 @@ function AdminProductCardEdit({ product, setIsOnEditMode }: Props) {
             onChange={(e) =>
               setEditedProduct({
                 ...editedProduct,
-                price: parseInt(e.target.value),
+                price: parseFloat(e.target.value),
               })
             }
           />
